+++ src/dune_rules/ordered_set_lang.mli
@@ -65,9 +65,6 @@ module Unexpanded : sig
 
   val has_standard : t -> bool
 
-  (** List of files needed to expand this set *)
-  val files : t -> f:(String_with_vars.t -> Path.t) -> Path.Set.t
-
   (** Expand [t] using with the given file contents. [file_contents] is a map
       from filenames to their parsed contents. Every [(:include fn)] in [t] is
       replaced by [Map.find files_contents fn]. Every element is converted to a
@@ -75,9 +72,8 @@ module Unexpanded : sig
   val expand :
        t
     -> dir:Path.t
-    -> files_contents:Dune_lang.Ast.t Path.Map.t
-    -> f:(String_with_vars.t -> Value.t list)
-    -> expanded
+    -> f:Value.t list Action_builder.t String_with_vars.expander
+    -> expanded Action_builder.t
 
   type position =
     | Pos