+++ test/blackbox-tests/test-cases/install-with-var.t/run.t
@@ -36,9 +36,9 @@ them could cause a dependency cycle (also, most of them make no sense in [dst] a
   > EOF
 
   $ dune build @install
-  File "dune", line 3, characters 27-42:
+  File "dune", line 3, characters 25-42:
   3 |   (files (foobar.txt as "%{env:FOO=foobar}/foo.txt"))
-                                 ^^^^^^^^^^^^^^^
+                               ^^^^^^^^^^^^^^^^^
   Error: %{env:..} isn't allowed in this position.
   [1]
 
@@ -65,9 +65,9 @@ extension of [src]:
   > EOF
 
   $ dune build @install
-  File "dune", line 3, characters 12-31:
+  File "dune", line 3, characters 10-31:
   3 |   (files (%{env:FOO=foobar.txt} as foo.txt))
-                  ^^^^^^^^^^^^^^^^^^^
+                ^^^^^^^^^^^^^^^^^^^^^
   Error: Because this file is installed in the 'bin' section, you cannot use
   the macro %{env:..} in its basename.
   [1]
@@ -104,9 +104,9 @@ Exe basename needs to be fully known if dst is missing though:
   > EOF
 
   $ dune build @install
-  File "dune", line 3, characters 11-26:
+  File "dune", line 3, characters 9-26:
   3 |   (files %{env:FOO=foobar}.txt)
-                 ^^^^^^^^^^^^^^^
+               ^^^^^^^^^^^^^^^^^
   Error: Because this file is installed in the 'bin' section, you cannot use
   the macro %{env:..} in its basename.
   [1]